Moved ContourMethodFiles to own folder
