+++ sensor_control/vision/symbols/src/ContourMethod/ContourMethod.cpp
