+++ sensor_control/vision/symbols/src/ContourMethod/FrameProc.cpp
