+++ sensor_control/vision/symbols/src/ContourMethod/ShapeDetector.h
