+++ sensor_control/vision/symbols/src/ContourMethod/ShapeFind.h
